added idx and frame_idx in Keyframes for better bookeeping + added list of keyframe_indices in Map
